SA pour trouver le minimum d'une fonction done

+++ src/main/java/SimulatedAnnealing/SAProblem.java
@@ -5,9 +5,13 @@ import java.util.ArrayList;
 public abstract class SAProblem {
 
     public abstract ArrayList<Object> getList();
+
     public abstract void printSolution(String s);
+
     public abstract SAProblem initialSolution();
+
     public abstract SAProblem transformSolution();
-    public abstract int objectiveFunction();
+
+    public abstract double objectiveFunction();
 
 }
